+++ src/OpenTelemetry.Resources.Azure/CHANGELOG.md
@@ -8,6 +8,9 @@
 * Add schema URL to resource detectors.
   ([#4775](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4775))
 
+* Fix `os.type` not being normalized to lowercase for Azure VMs.
+  ([#4775](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4775))
+
 ## 1.16.0-beta.1
 
 Released 2026-Jul-10